everything coompiles now

+++ eval_move.java
@@ -24,7 +24,7 @@ public class eval_move extends move
     {
         return value;
     }
-    public int compareTo(eval_move in)
+    public double compareTo(eval_move in)
     {
         return(this.get_value() - in.get_value());
     }
